Crate namada_vm

Virtual machine modules for running transactions and validity predicates.

Modules

host_env

Virtual machine’s host environment exposes functions that may be called from within a virtual machine.

memory

Virtual machine’s memory.

types

VM types can be passed between the host and guest via wasm linear memory.

Structs

HostRef

Reference to host environment data, to be used from wasm to implement host functions.

WasmCacheRoAccess

Restricted read-only access for dry-ran transactions

WasmCacheRwAccess

Regular read/write caches access

Enums

RoAccess

Read-only access to host data.

RwAccess

Read and write access to host data.

WasmValidationError

Traits

WasmCacheAccess

WASM Cache access level, used to limit dry-ran transactions to read-only cache access.

Functions

validate_untrusted_wasm

Validate an untrusted wasm code with restrictions that we place such code (e.g. transaction and validity predicates)

Type Aliases

RoHostRef

HostRef with read-only access.

RwHostRef

HostRef with read-write access.