namada_vm
Module host_env

Virtual machine’s host environment exposes functions that may be called from within a virtual machine.

Structs

TxCtx

A transaction’s host context

TxVmEnv

A transaction’s host environment

VpCtx

A validity predicate’s host context

VpVmEnv

A validity predicate’s host environment

Enums

Error

TxRuntimeError

These runtime errors will abort tx WASM execution immediately

Traits

VpEvaluator

A Validity predicate runner for calls from the vp_eval function.

Functions

tx_charge_gas

Add a gas cost incured in a transaction

tx_delete

Storage delete function exposed to the wasm VM Tx environment. The given key/value will be written as deleted to the write log.

tx_emit_event

Expose the functionality to emit events to the wasm VM’s Tx environment. An emitted event will land in the write log.

tx_get_block_epoch

Getting the block epoch function exposed to the wasm VM Tx environment. The epoch is that of the block to which the current transaction is being applied.

tx_get_block_header

Getting the block header function exposed to the wasm VM Tx environment.

tx_get_block_height

Getting the block height function exposed to the wasm VM Tx environment. The height is that of the block to which the current transaction is being applied.

tx_get_chain_id

Getting the chain ID function exposed to the wasm VM Tx environment.

tx_get_events

Expose the functionality to query events from the wasm VM’s Tx environment.

tx_get_native_token

Get the native token’s address

tx_get_pred_epochs

Get predecessor epochs function exposed to the wasm VM Tx environment.

tx_get_tx_index

Getting the transaction index function exposed to the wasm VM Tx environment. The index is that of the transaction being applied in the current block.

tx_has_key

Storage has_key function exposed to the wasm VM Tx environment. It will try to check the write log first and if no entry found then the storage.

tx_init_account

Initialize a new account established address.

tx_insert_verifier

Verifier insertion function exposed to the wasm VM Tx environment.

tx_iter_next

Storage prefix iterator next function exposed to the wasm VM Tx environment. It will try to read from the write log first and if no entry found then from the storage.

tx_iter_prefix

Storage prefix iterator function exposed to the wasm VM Tx environment. It will try to get an iterator from the storage and return the corresponding ID of the iterator, ordered by storage keys.

tx_log_string

Log a string from exposed to the wasm VM Tx environment. The message will be printed at the tracing::Level::INFO. This function is for development only.

tx_read

Storage read function exposed to the wasm VM Tx environment. It will try to read from the write log first and if no entry found then from the storage.

tx_read_temp

Read temporary value (not committed to storage) from the given key function exposed to the wasm VM Tx environment. It will try to read from the write log only.

tx_result_buffer

This function is a helper to handle the first step of reading var-len values from the host.

tx_set_commitment_sentinel

Set the sentinel for an invalid tx section commitment

tx_update_masp_note_commitment_tree

Appends the new note commitments to the tree in storage

tx_update_validity_predicate

Update a validity predicate function exposed to the wasm VM Tx environment

tx_verify_tx_section_signature

Verify a transaction signature

tx_write

Storage write function exposed to the wasm VM Tx environment. The given key/value will be written to the write log.

tx_write_temp

Temporary storage write function exposed to the wasm VM Tx environment. The given key/value will be written only to the write log. It will be never written to the storage.

tx_yield_value

Yield a byte array value from the guest.

vp_charge_gas

Called from VP wasm to request to use the given gas amount

vp_eval

Evaluate a validity predicate with the given input data.

vp_get_block_epoch

Getting the block epoch function exposed to the wasm VM VP environment. The epoch is that of the block to which the current transaction is being applied.

vp_get_block_header

Getting the block header function exposed to the wasm VM VP environment.

vp_get_block_height

Getting the block height function exposed to the wasm VM VP environment. The height is that of the block to which the current transaction is being applied.

vp_get_chain_id

Getting the chain ID function exposed to the wasm VM VP environment.

vp_get_events

Expose the functionality to query events from the wasm VM’s VP environment.

vp_get_native_token

Get the native token’s address

vp_get_pred_epochs

Get predecessor epochs function exposed to the wasm VM VP environment.

vp_get_tx_code_hash

Getting the transaction hash function exposed to the wasm VM VP environment.

vp_get_tx_index

Getting the block height function exposed to the wasm VM VP environment. The height is that of the block to which the current transaction is being applied.

vp_has_key_post

Storage has_key in posterior state (after tx execution) function exposed to the wasm VM VP environment. It will try to check the write log first and if no entry found then the storage.

vp_has_key_pre

Storage has_key in prior state (before tx execution) function exposed to the wasm VM VP environment. It will try to read from the storage.

vp_iter_next

Storage prefix iterator for prior or posterior state function exposed to the wasm VM VP environment.

vp_iter_prefix_post

Storage prefix iterator function for posterior state (after tx execution) exposed to the wasm VM VP environment.

vp_iter_prefix_pre

Storage prefix iterator function for prior state (before tx execution) exposed to the wasm VM VP environment.

vp_log_string

Log a string from exposed to the wasm VM VP environment. The message will be printed at the tracing::Level::INFO. This function is for development only.

vp_read_post

Storage read posterior state (after tx execution) function exposed to the wasm VM VP environment. It will try to read from the write log first and if no entry found then from the storage.

vp_read_pre

Storage read prior state (before tx execution) function exposed to the wasm VM VP environment. It will try to read from the storage.

vp_read_temp

Storage read temporary state (after tx execution) function exposed to the wasm VM VP environment. It will try to read from only the write log.

vp_result_buffer

This function is a helper to handle the first step of reading var-len values from the host.

vp_verify_tx_section_signature

Verify a transaction signature in the host environment for better performance

vp_yield_value

Yield a byte array value from the guest.

Type Aliases

Result

Result of a storage API call.

TxResult

Result of a tx host env fn call