namada_vm::host_env
Function vp_has_key_post

pub fn vp_has_key_post<MEM, D, H, EVAL, CA>(
    env: &mut VpVmEnv<MEM, D, H, EVAL, CA>,
    key_ptr: u64,
    key_len: u64,
) -> Result<i64>
where
    MEM: VmMemory,
    D: 'static + DB + for<'iter> DBIter<'iter>,
    H: 'static + StorageHasher,
    EVAL: VpEvaluator,
    CA: WasmCacheAccess,

Storage has_key in posterior state (after tx execution) function exposed to the wasm VM VP environment. It will try to check the write log first and if no entry found then the storage.