namada_vm::host_env
Function vp_result_buffer

pub fn vp_result_buffer<MEM, D, H, EVAL, CA>(
    env: &mut VpVmEnv<MEM, D, H, EVAL, CA>,
    result_ptr: u64,
) -> Result<()>
where
    MEM: VmMemory,
    D: 'static + DB + for<'iter> DBIter<'iter>,
    H: 'static + StorageHasher,
    EVAL: VpEvaluator,
    CA: WasmCacheAccess,

This function is a helper to handle the first step of reading var-len values from the host.

In cases where we’re reading a value from the host in the guest and we don’t know the byte size up-front, we have to read it in 2-steps. The first step reads the value into a result buffer and returns the size (if any) back to the guest, the second step reads the value from cache into a pre-allocated buffer with the obtained size.