namada_vm::host_env
Trait VpEvaluator

pub trait VpEvaluator {
    type Db: DB + for<'iter> DBIter<'iter>;
    type H: StorageHasher;
    type Eval: VpEvaluator;
    type CA: WasmCacheAccess;

    // Required method
    fn eval(
        &self,
        ctx: VpCtx<Self::Db, Self::H, Self::Eval, Self::CA>,
        vp_code_hash: Hash,
        input_data: BatchedTxRef<'_>,
    ) -> HostEnvResult;
}

A Validity predicate runner for calls from the vp_eval function.

Required Associated Types

type Db: DB + for<'iter> DBIter<'iter>

DB type

type H: StorageHasher

Storage hasher type

type Eval: VpEvaluator

Recursive VP evaluator type

type CA: WasmCacheAccess

WASM compilation cache access

Required Methods

fn eval( &self, ctx: VpCtx<Self::Db, Self::H, Self::Eval, Self::CA>, vp_code_hash: Hash, input_data: BatchedTxRef<'_>, ) -> HostEnvResult

Evaluate a given validity predicate code with the given input data. Currently, we can only evaluate VPs using WASM runner with WASM memory.

Invariant: Calling VpEvalRunner::eval from the VP is synchronous as it shares mutable access to the host context with the VP.