namada_vm::memory
Trait VmMemory

pub trait VmMemory:
    Clone
    + Send
    + Sync {
    type Error: Error + Sync + Send + 'static + Into<Error>;

    // Required methods
    fn read_bytes(
        &mut self,
        offset: u64,
        len: usize,
    ) -> Result<(Vec<u8>, Gas), Self::Error>;
    fn write_bytes(
        &mut self,
        offset: u64,
        bytes: impl AsRef<[u8]>,
    ) -> Result<Gas, Self::Error>;
    fn read_string(
        &mut self,
        offset: u64,
        len: usize,
    ) -> Result<(String, Gas), Self::Error>;
    fn write_string(
        &mut self,
        offset: u64,
        string: String,
    ) -> Result<Gas, Self::Error>;
}

Abstract representation of virtual machine’s memory.

Required Associated Types

type Error: Error + Sync + Send + 'static + Into<Error>

Error type for the methods’ results.

Required Methods

fn read_bytes( &mut self, offset: u64, len: usize, ) -> Result<(Vec<u8>, Gas), Self::Error>

Returns bytes read from memory together with the associated gas cost.

fn write_bytes( &mut self, offset: u64, bytes: impl AsRef<[u8]>, ) -> Result<Gas, Self::Error>

Write bytes to memory. Returns the gas cost.

fn read_string( &mut self, offset: u64, len: usize, ) -> Result<(String, Gas), Self::Error>

Returns string read from memory together with the associated gas cost.

fn write_string( &mut self, offset: u64, string: String, ) -> Result<Gas, Self::Error>

Write string to memory. Returns the gas cost.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.